requirement.txt
pandas
streamlit
python-dotenv
sqlalchemy
psycopg2-binary
matplotlib
flask